+++ integration-tests/tests/src/tests/sync/flexible.ts
@@ -494,7 +494,22 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(subs.state).to.equal(Realm.App.Sync.SubscriptionsState.Complete);
         });
 
-        it("is rejected if there is an error synchronising subscriptions", async function (this: RealmContext) {
+        // Should only pass if `"development_mode_enabled": true`
+        it("does not throw if querying a not explicitly queryable field (ONLY VALID IN DEV MODE)", async function (this: RealmContext) {
+          const { subs } = addSubscription(
+            this.realm,
+            this.realm.objects(FlexiblePersonSchema.name).filtered("nonQueryable == 'test'"),
+          );
+          expect(subs.state).to.equal(Realm.App.Sync.SubscriptionsState.Pending);
+
+          await subs.waitForSynchronization();
+
+          expect(subs.state).to.equal(Realm.App.Sync.SubscriptionsState.Complete);
+        });
+
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("is rejected if there is an error synchronising subscriptions", async function (this: RealmContext) {
           const { subs } = addSubscription(
             this.realm,
             this.realm.objects(FlexiblePersonSchema.name).filtered("nonQueryable == 'test'"),
@@ -508,7 +523,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(subs.state).to.equal(Realm.App.Sync.SubscriptionsState.Error);
         });
 
-        it("is rejected if subscriptions are already in an error state", async function (this: RealmContext) {
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("is rejected if subscriptions are already in an error state", async function (this: RealmContext) {
           const { subs } = addSubscription(
             this.realm,
             this.realm.objects(FlexiblePersonSchema.name).filtered("nonQueryable == 'test'"),
@@ -589,6 +606,20 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
         });
       });
 
+      describe("#length", function () {
+        it("returns the number of subscriptions", async function (this: RealmContext) {
+          expect(this.realm.subscriptions.length).to.equal(0);
+
+          const { subs, sub } = await addSubscriptionForPersonAndSync(this.realm);
+          expect(subs.length).to.equal(1);
+
+          await subs.update((mutableSubs) => {
+            mutableSubs.removeSubscription(sub);
+          });
+          expect(subs.length).to.equal(0);
+        });
+      });
+
       describe("array-like access", function () {
         async function addThreeSubscriptions(this: RealmContext) {
           addSubscriptionForPerson(this.realm);
@@ -604,13 +635,11 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(subs).to.have.length(0);
         });
 
-        it("accesses a SubscriptionSet using index operator", async function (this: RealmContext) {
-          const { subs } = await addThreeSubscriptions.call(this);
+        it("cannot access a SubscriptionSet using index operator", async function (this: RealmContext) {
+          const { subs } = await addSubscriptionForPersonAndSync(this.realm);
 
-          expect(subs).to.have.length(3);
-          expect(subs[0]).to.be.instanceOf(Realm.App.Sync.Subscription);
-          expect(subs[1]).to.be.instanceOf(Realm.App.Sync.Subscription);
-          expect(subs[2]).to.be.instanceOf(Realm.App.Sync.Subscription);
+          expect(subs).to.have.length(1);
+          expect(subs[0]).to.be.undefined;
         });
 
         it("spreads a SubscriptionSet using spread operator", async function (this: RealmContext) {
@@ -632,7 +661,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(numSubs).to.equal(3);
         });
 
-        it("iterates over a SubscriptionSet using 'Object.keys()'", async function (this: RealmContext) {
+        it("iterates over a SubscriptionSet using 'Object.keys()' (internal use)", async function (this: RealmContext) {
           const { subs } = await addThreeSubscriptions.call(this);
 
           // Object.keys() always returns an array of strings.
@@ -714,7 +743,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(subs.state).to.equal(Realm.App.Sync.SubscriptionsState.Complete);
         });
 
-        it("is Error if there is an error during synchronisation", async function (this: RealmContext) {
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("is Error if there is an error during synchronisation", async function (this: RealmContext) {
           await expect(
             addSubscriptionAndSync(
               this.realm,
@@ -727,7 +758,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(this.realm.subscriptions.state).to.equal(Realm.App.Sync.SubscriptionsState.Error);
         });
 
-        it("is Error if there are two errors in a row", async function (this: RealmContext) {
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("is Error if there are two errors in a row", async function (this: RealmContext) {
           await expect(
             addSubscriptionAndSync(
               this.realm,
@@ -786,7 +819,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(this.realm.subscriptions.error).to.be.null;
         });
 
-        it("contains the error message if there was an error synchronising subscriptions", async function (this: RealmContext) {
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("contains the error message if there was an error synchronising subscriptions", async function (this: RealmContext) {
           await expect(
             addSubscriptionAndSync(
               this.realm,
@@ -799,7 +834,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           );
         });
 
-        it("is null if there was an error but it was subsequently corrected", async function (this: RealmContext) {
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("is null if there was an error but it was subsequently corrected", async function (this: RealmContext) {
           await expect(
             addSubscriptionAndSync(
               this.realm,
@@ -821,7 +858,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           expect(this.realm.subscriptions.error).to.be.null;
         });
 
-        it("still contains the erroring subscription in the set if there was an error synchronising", async function (this: RealmContext) {
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("still contains the erroring subscription in the set if there was an error synchronising", async function (this: RealmContext) {
           await expect(
             addSubscriptionAndSync(
               this.realm,
@@ -943,7 +982,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
             expect(subs.state).to.equal(Realm.App.Sync.SubscriptionsState.Complete);
           });
 
-          it("returns a promise which is rejected if there is an error synchronising", async function (this: RealmContext) {
+          // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+          //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+          it.skip("returns a promise which is rejected if there is an error synchronising", async function (this: RealmContext) {
             const subs = this.realm.subscriptions;
 
             await expect(
@@ -980,14 +1021,15 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
 
           expect(subs).to.have.length(3);
 
-          expect(subs[0].queryString).to.equal("age < 10");
-          expect(subs[0].objectType).to.equal(FlexiblePersonSchema.name);
+          const subsCopy = [...subs];
+          expect(subsCopy[0].queryString).to.equal("age < 10");
+          expect(subsCopy[0].objectType).to.equal(FlexiblePersonSchema.name);
 
-          expect(subs[1].queryString).to.equal("age > 20");
-          expect(subs[1].objectType).to.equal(FlexiblePersonSchema.name);
+          expect(subsCopy[1].queryString).to.equal("age > 20");
+          expect(subsCopy[1].objectType).to.equal(FlexiblePersonSchema.name);
 
-          expect(subs[2].queryString).to.equal("age > 30");
-          expect(subs[2].objectType).to.equal(DogSchema.name);
+          expect(subsCopy[2].queryString).to.equal("age > 30");
+          expect(subsCopy[2].objectType).to.equal(DogSchema.name);
         });
 
         it("handles multiple updates in multiple batches", async function (this: RealmContext) {
@@ -1005,17 +1047,20 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
 
           expect(subs).to.have.length(3);
 
-          expect(subs[0].queryString).to.equal("age < 10");
-          expect(subs[0].objectType).to.equal(FlexiblePersonSchema.name);
+          const subsCopy = [...subs];
+          expect(subsCopy[0].queryString).to.equal("age < 10");
+          expect(subsCopy[0].objectType).to.equal(FlexiblePersonSchema.name);
 
-          expect(subs[1].queryString).to.equal("age > 20");
-          expect(subs[1].objectType).to.equal(FlexiblePersonSchema.name);
+          expect(subsCopy[1].queryString).to.equal("age > 20");
+          expect(subsCopy[1].objectType).to.equal(FlexiblePersonSchema.name);
 
-          expect(subs[2].queryString).to.equal("age > 30");
-          expect(subs[2].objectType).to.equal(DogSchema.name);
+          expect(subsCopy[2].queryString).to.equal("age > 30");
+          expect(subsCopy[2].objectType).to.equal(DogSchema.name);
         });
 
-        it("still applies all updates in a batch if one errors", async function (this: RealmContext) {
+        // TODO: Enable test when we can find another way of triggering a `SubscriptionsState.Error`.
+        //       (This is due to non queryable fields now being queryable since BaaS automatically adds them when in Dev Mode)
+        it.skip("still applies all updates in a batch if one errors", async function (this: RealmContext) {
           const { subs } = await addSubscriptionForPersonAndSync(this.realm);
 
           await expect(
@@ -1070,8 +1115,9 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           const { subs } = await addSubscriptionForPersonAndSync(this.realm, { name: "test2" });
 
           expect(subs).to.have.lengthOf(2);
-          expect(subs[0].name).to.equal("test1");
-          expect(subs[1].name).to.equal("test2");
+          const subsCopy = [...subs];
+          expect(subsCopy[0].name).to.equal("test1");
+          expect(subsCopy[1].name).to.equal("test2");
         });
 
         it("does not add a second identical subscription with the same name", async function (this: RealmContext) {
@@ -1109,7 +1155,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           );
 
           expect(subs).to.have.lengthOf(1);
-          expect(subs[0].queryString).to.equal("age > 10");
+          expect([...subs][0].queryString).to.equal("age > 10");
         });
 
         it("allows an anonymous and a named subscription for the same query to exist", async function (this: RealmContext) {
@@ -1126,7 +1172,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           );
 
           expect(subs).to.have.lengthOf(2);
-          expect(subs[1].id).to.not.equal(sub.id);
+          expect([...subs][1].id).to.not.equal(sub.id);
         });
 
         it("if a subscription with the same query is added, properties of both the old and new reference can be accessed", async function (this: RealmContext) {
@@ -1209,7 +1255,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           });
 
           expect(subs).to.have.length(1);
-          expect(subs[0].queryString).to.equal("age > 10");
+          expect([...subs][0].queryString).to.equal("age > 10");
         });
       });
 
@@ -1238,7 +1284,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           });
 
           expect(subs).to.have.length(1);
-          expect(subs[0].queryString).to.equal("age > 10");
+          expect([...subs][0].queryString).to.equal("age > 10");
         });
 
         it("removes multiple subscriptions", async function (this: RealmContext) {
@@ -1256,7 +1302,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           });
 
           expect(subs).to.have.length(1);
-          expect(subs[0].queryString).to.equal("age > 15");
+          expect([...subs][0].queryString).to.equal("age > 15");
         });
       });
 
@@ -1292,7 +1338,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           });
 
           expect(subs).to.have.length(1);
-          expect(subs[0].queryString).to.equal("age > 10");
+          expect([...subs][0].queryString).to.equal("age > 10");
         });
 
         it("if a subscription with the same query is added, the old reference can be removed", async function (this: RealmContext) {
@@ -1364,7 +1410,7 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           });
 
           expect(subs).to.have.length(1);
-          expect(subs[0].objectType).to.equal(DogSchema.name);
+          expect([...subs][0].objectType).to.equal(DogSchema.name);
         });
       });