+++ packages/bindgen/realm-bindgen.ts
@@ -1,3 +1,3 @@
-#!/usr/bin/env -S tsx --
+#!/usr/bin/env -S npx tsx --
 import { program } from "./src/program";
 program.parse(process.argv);