+++ packages/realm-app-importer/realm-app-importer.ts
@@ -1,2 +1,2 @@
-#!/usr/bin/env -S tsx --
+#!/usr/bin/env -S npx tsx --
 import "./src/cli";